% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-frame.R
\name{apply_check}
\alias{apply_check}
\title{Run a sample frame's check rules}
\usage{
apply_check(df, sfw, verbose = TRUE)
}
\arguments{
\item{df}{A data frame.}

\item{sfw}{A sample frame.}

\item{verbose}{Print the results.}
}
\value{
Invisibly, a data frame of check results.
}
\description{
Run a sample frame's check rules
}
