% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmf.R
\name{bmf_states}
\alias{bmf_states}
\title{Return the published BMF state mart codes}
\usage{
bmf_states(format = c("csv", "parquet"))
}
\arguments{
\item{format}{\code{"csv"} (default) or \code{"parquet"}.}
}
\value{
A named character vector of state mart URLs, named by postal code.
}
\description{
Return the published BMF state mart codes
}
