% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{efile_version}
\alias{efile_version}
\title{Current default efile release}
\usage{
efile_version()
}
\value{
The version string used by \code{\link[=data_source]{data_source()}} when none is supplied.
}
\description{
Current default efile release
}
