% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{data_source}
\alias{data_source}
\title{Create an efile source configuration}
\usage{
data_source(root = .EFILE_ROOT, aliases = .EFILE_ALIASES)
}
\arguments{
\item{root}{Base URL or local directory containing table-year CSV files.}

\item{aliases}{Named character vector mapping short aliases to table names.}
}
\value{
An \code{data_source} object.
}
\description{
Create an efile source configuration
}
