% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{table_catalog}
\alias{table_catalog}
\title{Return the built-in efile table aliases}
\usage{
table_catalog()
}
\value{
A data frame with alias, canonical table, and cardinality.
}
\description{
Return the built-in efile table aliases
}
